fix: fire the `change` event for `lablup-slider`
The session launcher relies on the `change` event of `lablup-slider` to
update the values of the resources. However, the `lablup-slider` was not
firing the `change` event when the `mwc-textfield` is updated. Instead,
it actually fires the `changed` and `value-change` events and that's why
the session launcher does not recognize the updated resources when user
changed the values with the `mwc-textfield`s.

Also, `mwc-slider` fires its own `change` event. That explains why the
session launcher recognizes the resource changes when user updates them
with the sliders.

This commit fixes the logic by:
* When the `mwc-slider` is updated, `lablup-slider` does not fire the
  `change` event since it already is fired by the slider, implicitly.
* When the `mwc-textfield` is updated, `lablup-slider` explicitly fires
  the `change` event since the textfield does not fire any event upon
  changes.

The `mwc-slider` has another problem that the knob is not udpated unless
we manually update the component with the `layout()` method. To resolve
this issue, I introduced the `intersectionObserver` to detect when the
`lablup-slider` is visible to the user, and manually update the slider's
knob by calling the `layout` method every time it turns to visible.

TODO: There is no `marker` property in the `mwc-slider`, so it is
      useless for now. Since it is not essential, I haven't touch the code
      related to the markers.