feat: hide vfolder sharing buttons when invite-others permission is disabled (#1642)
* add: composite graphQL query to receive allowed_vfolder_hosts at once

* refactor: hide vfolder sharing related icon-button when `invite-other` permission is not allowed in any policy associated to the current user

* refactor: getAllowedVFolderHostsByCurrentUserInfo

* refactor: query installed images only in checking filebrowser supported images

The previous logic fetches every image including not-installed ones, and
then filter installed ones. This takes quite some time if there are
several hundreds of registered images. Since the image list query API
already supports fetching the installed ones only, we don't need to
fetch not-installed images.

* fix: remove console error after creating a data folder

The error below was raised to update the auto mount folder tab since
the `this.folderListGrid` is `null` in the page:

```
Unhandled Promise Rejection: TypeError: null is not an object
(evaluating 'this.folderListGrid.clearCache')
```

* refactor: more responsiveness using async calls in viewStateChanged

* refactor: add element ID for preferred app port config box in app launcher

---------

Co-authored-by: Jonghyun Park <jpark@lablup.com>