fix: non-castable active property on components
For a boolean property to be configurable from an attribute, it must
default to false. If it defaults to true, you cannot set it to false
from markup, since the presence of the attribute, with or without a
value, equates to true. This is the standard behavior for attributes in
the web platform. (from https://lit.dev/docs/components/properties/ )