fix: include cancel handler in ServiceLauncherModal (#2379)
### TL;DR

This pull request introduces a change to the `onCancel` handling in the `ServiceLauncherModal` to resolve the issue of the modify service modal popping up when clicking the start service button.

### What changed?
In the `ServiceListPage`, the `onRequestClose` method is called to set the `editingModelService` to null and remove the `onCancel` prop. 

In the `ServiceLauncherModal`, the `onCancel` behavior has been changed to call `onRequestClose` when triggered.

### How to test?

1.  Navigate to the serving page.
2. Click the gear icon in the Control column and close the modal with the close button at the bottom (not the x button).
3. Click Create new service button.
4. The create modal should appear.


---

<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->
resolves https://github.com/lablup/backend.ai-webui/issues/2378

**Checklist:** (if applicable)

- [x] Mention to the original issue
- [ ] Documentation
- [ ] Minium required manager version
- [x] Specific setting for review (eg., KB link, endpoint or how to setup) : 10.100.64.15
- [x] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after