Omit `onOk` from `BAIModalProps` instead of `onOK` (#2384)
### TL;DR

This change fixes the inconsistency in the props of EndpointTokenGenerationModal.tsx.

### What changed?

Property 'onOK' is corrected to 'onOk' in EndpointTokenGenerationModal.tsx.

### How to test?

Pull the changes, and verify the property in the EndpointTokenGenerationModal.tsx component.

### Why make this change?

The change was necessary to align with the correct property name and prevent potential prop related issues.

---

<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->

**Checklist:** (if applicable)

- [ ] Mention to the original issue
- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after