fix: reload behavior of BAIErrorBoundary for Electron (#2397)
### TL;DR

This pull request modifies the BAIErrorBoundary component's refresh functionality to be compatible with the Electron environment.

### What changed?

The PR updates the onClick handler of the refresh button in the Error View. When the refresh button is clicked, the application checks whether it's running in an Electron environment. If so, it refers to the 'electronInitialHref' for reloading; otherwise, it uses the standard location.reload().

### How to test?

You can test this by triggering an error on the page and clicking the refresh button, in both web browser and Electron environments.
- Example of triggering an error: 
  - Build electron app and launch
  - turn off the network of you machine
  - move the "Service" page and click "Start service"
  - You can see the Error Boundary fallback view

### Why make this change?

This change ensures that the correct refresh method is called depending upon the environment, thereby preventing potential errors or undesired behavior in Electron-based applications.

---

<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->

**Checklist:** (if applicable)

- [ ] Mention to the original issue
- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after