feat: provide sorted resource presets to NeoSessionLauncher when presets are modifed (#2435)
### Feature 
When the Resource Preset is modified, the item is moved to the end of the list. This makes it difficult for users to find the resource preset, so we provide sorting functionality by name for resource presets to improve the user experience.

<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->

**Checklist:** (if applicable)

- [ ] Mention to the original issue
- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after