fix: temporally disable session commit to tar file in commit session dialog (#2433)
### TL;DR

This PR temporally disables commit session to `tar.gz` file in agent node, since there's no way to manage it as images in My Environment page.

### What changed?
Back then, commit session dialog has two feature, 1) Simply commit session to agent node and 2) Commit session to agent node and convert it to executable image. We disabled the first feature in this PR.

### How to test?
1. Create a session.
2. Click container commit icon in Controls column.
3. Dialog pops up. 
4. See commit button is disabled.
5. See the commit button changed to clickable status when you click the checkbox and input title of customized image at the input field.

### Why make this change?
For now we don't have any kind of management UI to limit and delete user-commit file, which may bring shortage of storage. Therefore It should be blocked.

---

<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->

**Checklist:** (if applicable)

- [ ] Mention to the original issue
- [ ] Documentation
- [ ] Minimum required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [x] Test case(s) to demonstrate the difference of before/after