fix: improve `message`'s i18n and duration for model services deletion  (#2439)
### What I Fixed
Changed the wording that appears when a model service is terminated. Currently, when you request a service to be terminated, it doesn't happen immediately, but takes a while because it depends on some dependencies, so I've changed the wording to be more request-driven.
### Feature
|before|after|
|---|---|
|![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/lSyr8xXz1wdXALkJKzVx/bf3e81d1-9be2-4110-a242-d47f823d7b69.png)|![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/lSyr8xXz1wdXALkJKzVx/8d846644-e88c-4a43-a8a7-716ca96e625f.png)|
<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->

**Checklist:** (if applicable)

- [ ] Mention to the original issue
- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after