feature: hide user resource policy based on manager version (#2453)
### TL;DR

This PR updates the resource policy page.

### What changed?

In versions 23.09.10 and earlier, it is not suitable to create and modify user resource policies. So this PR will hide user resource policy if manager version < 23.09.10.

---

<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->

**Checklist:** (if applicable)

- [ ] Mention to the original issue
- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after