fix: run form validation after changing preset (#2480)
This pull request ensures that when the 'resource' field is validated when preset is changed, the validation is conducted recursively. It addresses an issue where non-recursive validation could have missed nested fields under 'resource'. The change involves modifying the validateFields method to include the 'recursive: true' option. This update impacts the ResourceAllocationFormItems component in the React codebase.

---

<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->

**Checklist:** (if applicable)

- [ ] Mention to the original issue
- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after