feat: Lucide icon with custom svgs (#2458)
### TL;DR

Add `lucide-react` dependency to the project.

This PR change a  only one icon for testing and improving UI.
| Before | After |
| -- | -- |
|  ![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XqC2uNFuj0wg8I60sMUh/92a511a7-b276-430f-93f6-9a010ffc47ae.png) |  ![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XqC2uNFuj0wg8I60sMUh/d8c0ed05-b34e-4ba7-9b42-05bb71dc92a3.png) |

FYI, only the icons that are imported into your project are included in the final bundle.

### What changed?

This PR adds the `lucide-react` library (version ^0.383.0) as a dependency in both `package.json` and `package-lock.json` files. The `lucide-react` library provides a set of SVG icons for React.

### How to test?

1. Run `npm install` to update your local dependencies.
2. Ensure there are no installation errors.
3. Verify that the `lucide-react` library is correctly listed in your `node_modules` directory.
4. Import and use any icon from `lucide-react` in your React components to ensure it's working as expected.

### Why make this change?

The `lucide-react` library adds a collection of essential React icons, enhancing our project's UI consistency and reducing the need to import individual icons from different sources.

---

<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->

**Checklist:** (if applicable)

- [ ] Mention to the original issue
- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after