fix: 400 error occurs when refreshing container logs (#2536)
follows #2524

### TL;DR

This pull request updates the `BackendAISessionList` component to support retrieving logs specific to a selected kernel.

### What changed?

The `get_logs` method of `backendaiclient` has been modified to accept an additional parameter, `kernelId`. If a kernel is selected, logs for that specific kernel will be retrieved. If no kernel is selected, all logs will be retrieved as before. This change affects two places in the `backend-ai-session-list.ts` file where `get_logs` is called.

### How to test?

1. In the UI, select a specific kernel.
2. Trigger a log retrieval action.
3. Verify that the logs retrieved are specific to the selected kernel.
4. Unselect the kernel and trigger log retrieval again.
5. Verify that all logs are retrieved when no kernel is selected.

### Why make this change?

The motivation behind this change is to provide users with the ability to filter logs by specific kernels, which can be useful for debugging and monitoring purposes.

---

<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->

**Checklist:** (if applicable)

- [x] Mention to the original issue
- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after