feat: re-render once when `backend-ai-page` is inactive (#2581)
### TL;DR

Added a 'requiresUpdateWhenInactive' flag to ensure the `backend-ai-page` component updates at least once when it becomes inactive.

### What changed?
The BackendAIPage component now includes a 'requiresUpdateWhenInactive' flag. This ensures that an update is triggered once when the component goes inactive. Additionally, updated the BackendAISessionList component to render the session list grid dynamically based on the 'active' state.

### How to test?
1. Navigate to the page using the BackendAIPage component.
2. Visit session list and move to other page to trigger active/inactive state and observe if the session list updates accordingly.
- When inactive, the `ReservationTimeCounter` React component in the session list should be unmounted. You can ensure it to check
 whether the [`useIntervalValue` hook is cleared](https://github.com/lablup/backend.ai-webui/blob/15d929f4891b1d56c0e13ec819d979e8b638fb14/react/src/hooks/useIntervalValue.tsx#L23-L24
).

### Why make this change?
This change addresses an issue where the session list would not update properly when becoming inactive. Ensuring an update is triggered helps maintain the consistency and accuracy of the displayed information.

---

<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->

**Checklist:** (if applicable)

- [ ] Mention to the original issue
- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after