fix: typo error (#2593)
### TL;DR
Fixed a typo in the Korean localization file `ko.json`.

### What changed?
Corrected the word "탐석" to "탐색" in the key "FindingSessionTemplate".

### How to test?
No specific testing required, but you can verify the change by checking the affected key in `ko.json` file for correct spelling.

### Why make this change?
This fixes a minor typo to ensure accurate translations in the Korean localization.

---

<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->

**Checklist:** (if applicable)

- [ ] Mention to the original issue
- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after