fix: set notification's description field to empty string (#2601)
### Why did I make the change?
Because the `backend-ai-session-list` shares a single description stored variable with the notification, it is displaying the description as it is if it was stored by another action.

So, in the part that shows the notifications, if you don't want a description, you have to explicitly put in an empty string value.

<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->

**Checklist:** (if applicable)

- [ ] Mention to the original issue
- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after