feat: improve wsproxy address validation (#2602)
resolves https://github.com/lablup/backend.ai-webui/issues/2597

This pull request addresses a cosmetic issue by fixing a typo in the variable naming convention for resourceGroupWSProxyAddressInput. It also includes enhancements to the styling of the input fields by using the token value for the text label color in disabled and secondary states. Moreover, the validation pattern and auto-validation for wsproxy address input have been added to ensure correct URL format.

### How to test
1. go to resources - resource group tab
2. click the setting icon
3. check validation of the wsproxy address.

---

<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->

**Checklist:** (if applicable)

- [x] Mention to the original issue
- [ ] Documentation
- [ ] Minium required manager version
- [x] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after