feat: add confirm dialog before moving to trash bin tab (#2610)
### TL;DR

Changed the action label for moving a folder to trash in various languages and added a confirmation dialog for the delete(move to the trash bin) button.

### What changed?

- Updated the label "Move to trash bin" to "Move to trash" in translation files for multiple languages.
- Added a description asking for confirmation before moving a folder to the trash in the UI.

### How to test?

1. Attempt to move a folder to the trash.
2. Confirm the appearance of the confirmation dialog with the updated text.
3. Validate the folder is moved to the trash upon confirming the action.

### Why make this change?

To improve user experience by providing a clear and concise action description and ensuring users confirm their intent before moving a folder to the trash.

### Screenshot

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/2HueYSdFvL8pOB5mgrUQ/5f8869ec-5471-495c-849c-b03004169100.png)

---

<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->

**Checklist:** (if applicable)

- [ ] Mention to the original issue
- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after