feat: persist current Tab on the Environment page using search param (#2587)
<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->

### This PR resolves [#2586 ](https://github.com/lablup/backend.ai-webui/issues/2586) issue
### TL;DR

manage tab as URL

###  Why make this change?

since we didn't manage the tab as a URL, it doesn't persist on refresh.

### What changed?

modify to use useQueryParam to manage tabs

` curTabKey === 'registryList'`  This is unnecessary code, so I deleted it
 because if curTabKey is not registryList, the parent component's display is none
```
 curTabKey === 'registryList' ? (
            <Suspense>
              <ContainerRegistryList />
            </Suspense>
          ) : null
```
### How to test?

1. enter the environment page
2. click preset or registry and check it's working well
3. refresh and check the tab is preserved