feature: set `architecture` as a second-order field in Image Select (#2666)
### TL;DR

Updated the image sorting logic in the ImageEnvironmentSelectFormItems component.

### What changed?

Modified the image sorting algorithm in the ImageEnvironmentSelectFormItems component. The sorting now considers both version and architecture. If versions are equal, it falls back to comparing architectures using the newly imported `localeCompare` function.

### Why make this change?

It provides a more consistent and user-friendly experience, especially when dealing with images that have the same version but different architectures.
**Additionally, it's very helpful when writing e2e test code.**